+++ Changelog.md
@@ -1,7 +1,10 @@
 # CHANGELOG
 All notable changes to this project will be documented here.
 
-## [unreleased]
+## [v1.10.0]
+- Updated development docker image to connect to the development MarkUs docker image (#238)
+- Removed Tasty-Stats as a dependency for the haskell tester and added our own ingredient instead to collect stats (#259)
+- Updated/improved the interface between the autotester and clients that use it (#257)
 
 ## [1.9.0]
 - allow tests to write to existing subdirectories but not overwrite existing test script files (#237).